refactor: change SalesSummaryResponse type

+++ src/main/kotlin/com/kroffle/knitting/controller/handler/knitter/MyselfHandler.kt
@@ -40,7 +40,7 @@ class MyselfHandler(
                 ResponseHelper
                     .makeJsonResponse(
                         SalesSummaryResponse(
-                            numberOfProductsOnSales = it.toInt(),
+                            numberOfProductsOnSales = it,
                             numberOfProductsSold = 0,
                         )
                     )
